Move config to seperate file
